MNT: move areadetector cameras into separate file
