+++ ophyd/controls/areadetector/__init__.py
@@ -4,3 +4,8 @@ import logging
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
+
+from .base import *
+from .cam import *
+from .detectors import *
+from .plugins import *
